Add basic resource system

+++ src/server/performers/damagingperformer.cpp
@@ -10,8 +10,8 @@
 namespace server::damaging_performer {
 
 void damage(std::shared_ptr<core::GameWorld> world, std::shared_ptr<core::Object> object,
-            std::shared_ptr<core::Damaging> damaging, double timeDelta) {
-    double delayLeft = damaging->getCurrentDelay() - timeDelta;
+            std::shared_ptr<core::Damaging> damaging, int timeDelta) {
+    int delayLeft = damaging->getCurrentDelay() - timeDelta;
     if (delayLeft > 0) {
         damaging->setCurrentDelay(delayLeft);
         return;
